Refactor with main_core
Splitting Calibration dialog into many classes
