+++ python/AiWarmachine/constants.py
@@ -25,6 +25,9 @@ IS_LINUX = sys.platform.startswith('linux')
 
 CRITERIA = (cv.TERM_CRITERIA_EPS + cv.TERM_CRITERIA_MAX_ITER, 30, 0.001)
 
+DEFAULT_TICKS_PER_SECONDS = 30.0
+DEFAULT_QR_DETECTION_PER_SECONDS = 4.0
+
 DEFAULT_CAPTURE_API = cv.CAP_V4L2 if IS_LINUX else cv.CAP_DSHOW
 
 DEFAULT_CAPTURE_WIDTH = 1920
@@ -115,3 +118,5 @@ TABLE_CORNERS_INDEX_TR = 2
 TABLE_CORNERS_INDEX_BR = 3
 
 TABLE_CORNERS_DRAWING_ORDER = [TABLE_CORNERS_INDEX_BL, TABLE_CORNERS_INDEX_TL, TABLE_CORNERS_INDEX_TR, TABLE_CORNERS_INDEX_BR]
+
+QR_CENTER_EPISLON = 4
