+++ python/AiWarmachine/viewport_label.py
@@ -1,3 +1,19 @@
+#
+# This file is part of the AiWarmachine distribution (https://github.com/madesjardins/AiWarmachine).
+# Copyright (c) 2023 Marc-Antoine Desjardins.
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, version 3.
+#
+# This program is distributed in the hope that it will be useful, but
+# WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
+# General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program. If not, see <http://www.gnu.org/licenses/>.
+#
 """Viewport label widget."""
 
 from PyQt6 import QtCore, QtWidgets
